made a license button and window albeit with a poor implementation (decided to go with Apache nstead of gpl {always thought my frst program would be gpl}) and Finally have silent running, but:
(1) it still throws an error about splashscreen1 being called while the application is trying to close.

(2) the application does blind silent: meanng it just attempts to install any command-line argument passed to it.

(3) no result is printed to command-line output; need to implement a text output for silent running for when it runs from cmd, ps or any other terminal-like application.